requirements.txt
streamlit
Pillow
pandas
requests
opencv-python
